Салон: Уровень CO2 в запросе к БС-СМ умножается на 10

+++ source/CustomModbusSlave.Es2gClimatic.InteriorApp/BsSm/Contracts/IBsSmAndKsm1DataCommand32Request.cs
@@ -18,7 +18,7 @@
 		int Fault5 { get; }
 
 		int ReserveOrTemperatureAverageValue { get; }
-		int Co2LevelInCurrentSegment { get; }
+		double Co2LevelInCurrentSegment { get; }
 		int Reserve25 { get; }
 	}
 }
